--- a/.gitattributes
+++ b/.gitattributes
@@ -33,4 +33,4 @@
 *.woff binary
 *.woff2 binary
 *.pyc binary
-*.pdf binary 
+*.pdf binary