--- a/.gitignore
+++ b/.gitignore
@@ -68,3 +68,12 @@ temp/
 reports/mutation/mutation.html
 reports/coverage/
 reports/performance/
+
+
+
+.nx/cache
+.nx/workspace-data
+.cursor/rules/nx-rules.mdc
+.github/instructions/nx.instructions.md
+# Test result comparison files
+test-results/