--- a/.husky/pre-push
+++ b/.husky/pre-push
@@ -1,0 +1,29 @@
+#!/usr/bin/env sh
+. "$(dirname -- "$0")/_/husky.sh"
+
+# Quick validation before push
+echo "🔍 Running pre-push validation..."
+
+# Check for uncommitted changes
+if ! git diff --quiet || ! git diff --cached --quiet; then
+  echo "❌ Uncommitted changes detected. Please commit or stash changes before pushing."
+  exit 1
+fi
+
+# Run quick lint and type checks
+echo "📝 Running lint and type checks..."
+pnpm ai:quick || {
+  echo "❌ Lint or type check failed. Run 'pnpm lint:fix' to fix issues."
+  exit 1
+}
+
+# Quick config validation (only utils tests for speed)
+echo "🧪 Validating test configurations..."
+pnpm test:config:quick || {
+  echo "❌ Test configuration mismatch detected!"
+  echo "This could cause mutation testing to fail in CI."
+  echo "Run 'pnpm test:config:verify' for full validation."
+  exit 1
+}
+
+echo "✅ Pre-push validation passed!"