… .actrc
… .cursor/rules/default.mdc
… .github/workflows/test-config-validation.yml
… .gitleaksignore
… .husky/pre-commit
… .nvmrc
… .prettierignore
… AGENTS.md
… CLAUDE.md
… apps/backend-api/.nvmrc
… apps/backend-api/.prettierignore
… apps/backend-api/openapi.json
… apps/backend-api/package.json
… apps/backend-api/reports/mutation/mutation.html
… apps/backend-api/test/tsconfig.json
… apps/backend-api/tsconfig.json
… docs/DEVELOPMENT.md
… docs/ENTERPRISE_ROADMAP.md
… docs/LOCAL_VALIDATION_GUIDE.md
… docs/TESTING_COOKBOOK.md
… eslint.config.old.js
… nx.json
… package.json
… packages/config/.prettierignore
… packages/config/.tsbuildinfo
… packages/config/package.json
… packages/config/test/tsconfig.json
… packages/config/tsconfig.json
… packages/config/vitest.config.ts
… packages/types/.tsbuildinfo